requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pydantic
httpx
google-generativeai
